+++ functions.php
@@ -7,5 +7,5 @@
 
 use MacrosBySara\Theme_Init;
 
-require_once get_template_directory() . '/inc/class-theme-init.php';
+require_once get_stylesheet_directory() . '/inc/theme/class-theme-init.php';
 new Theme_Init();