+++ changelog.md
@@ -1,5 +1,7 @@
 # Changelog
 
-## 1.0.0 - [August 24, 2024]
+## 1.0.1 - [August 24, 2024]
 
 -   Init theme!
+-   Fix stylesheet path errors
+-   Add more-intense speculative loading